config: remove unneeded path field
As well as passing the key value pair, the config machinery also passes
a "struct key_value_info" containing information about the source of
that pair to the config parsing callback. Accessing the "path" field
of this struct from a callback passed to repo_config() results in a
use-after-free. This happens because repo_config() first populates a
configset by calling config_with_options() and then iterates over the
configset with the callback passed by the caller. When the configset
is constructed it takes a shallow copy of the "struct key_value_info"
for each config pair this leads to the use-after-free as the "path"
member is freed before config_with_options() returns.

We could fix this by interning the "path" field as we do for "filename"
but the "path" field is not actually needed. It is a copy of the same
field from "struct config_source". That field was added in d14d42440d8
(config: disallow relative include paths from blobs, 2014-02-19)
to distinguish relative include directives in files from those in
blobs. However since 1b8132d99d8 (i18n: config: unfold error messages
marked for translation, 2016-07-28) we can differentiate these by
looking at the config origin. So let's remove the "path" members from
"struct config_source" and "struct key_value_info" and instead use a
combination of "filename" and "origin_type" to determine the absolute
path of relative includes.

Signed-off-by: Phillip Wood <phillip.wood@dunelm.org.uk>