t1403: remove dependency on GIT_TEST_DEFAULT_INITIAL_BRANCH_NAME
Some of the tests use GIT_TEST_DEFAULT_INITIAL_BRANCH_NAME when
checking if a ref exists. However, as that variable is always set to
"main" at the start of the test file, we can hard code the name in the
tests in preparation for removing GIT_TEST_DEFAULT_INITIAL_BRANCH_NAME
in Git 3.0.

Signed-off-by: Phillip Wood <phillip.wood@dunelm.org.uk>