replay: drop commits that become empty
If the changes in a commit being replayed are already in the branch
that the commits are being replayed onto then "git replay" creates an
empty commit. This is confusing because the commit message no longer
matches the contents of the commit. Drop the commit instead. Commits
that start off empty are not dropped. This matches the behavior of
"git rebase --reapply-cherry-pick --empty=drop" and "git cherry-pick
--empty-drop". If a branch points to a commit that is dropped it will
be updated to point to the last commit that was not dropped. This can
been seen in the new test where "topic1" is updated to point to the
rebased "C" as "F" is dropped because it is already upstream. While
this is a breaking change "git replay" is marked as experimental to
allow improvements like this that change the behavior.

Signed-off-by: Phillip Wood <phillip.wood@dunelm.org.uk>